Main bulk of the Spirit ability implementation
Spirit ability is like... think of Berserk. Has very similar conditions like activating before berries but can also activate during multistrike moves similarly to berries unlike Berserk. If any of the benefits are successful, either the attack or special attack boost or the heal, then that pokemon's Spirit ability will be expended for the rest of the battle.